refactor(frontend): move line utilities and auto jump utilities into separate files
