+++ app/frontend/src/Overlay/index.js
@@ -8,7 +8,7 @@ import ThemeLoader from './ThemeLoader'
 import { SettingsContext, StatusContext } from '../lib/contexts'
 import { LANGUAGES } from '../lib/consts'
 import { useTranslations, useCurrentLine } from '../lib/hooks'
-import { customiseLine, getTransliterators } from '../lib/utils'
+import { customiseLine, getTransliterators } from '../lib/line'
 
 import './index.css'
