+++ app/frontend/src/Presenter/Line.js
@@ -5,7 +5,7 @@ import { CSSTransition, TransitionGroup } from 'react-transition-group'
 import classNames from 'classnames'
 import { countSyllables, toSyllabicSymbols } from 'gurmukhi-utils'
 
-import { partitionLine, classifyWords } from '../lib/utils'
+import { partitionLine, classifyWords } from '../lib/line'
 import { DEFAULT_OPTIONS } from '../lib/options'
 import { LANGUAGES, LANGUAGE_NAMES, TRANSLATION_ORDER, TRANSLITERATION_ORDER } from '../lib/consts'
