+++ app/frontend/src/shared/NavigatorHotkeys.js
@@ -2,7 +2,8 @@ import React, { useContext, useEffect, useCallback } from 'react'
 import { node, bool, shape, instanceOf } from 'prop-types'
 import { noop } from 'lodash'
 
-import { getJumpLines, findLineIndex } from '../lib/utils'
+import { findLineIndex } from '../lib/line'
+import { getJumpLines } from '../lib/auto-jump'
 import controller from '../lib/controller'
 import { NAVIGATOR_SHORTCUTS, LINE_HOTKEYS } from '../lib/keyMap'
 import { ContentContext, HistoryContext, SettingsContext } from '../lib/contexts'
